Move Schemes into GDALPath
Its used exclusivly there and avoids introducing another top level class.
Potentially it could even be flatterend if not for somewhat confusing prefexing.
